refactor(android): rename imports to be com.shabados.app
