+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.shabados.mobile",
+    package = "com.shabados.app",
 )
 
 android_resource(
     name = "res",
-    package = "com.shabados.mobile",
+    package = "com.shabados.app",
     res = "src/main/res",
 )
