+++ android/app/src/main/java/com/shabados/app/MainActivity.java
@@ -1,4 +1,4 @@
-package com.shabados.mobile;
+package com.shabados.app;
 
 import com.facebook.react.ReactActivity;
